+++ SocialNetwork/client/src/app/components/profile-page/user-page/user-page.component.html
@@ -1,6 +1,6 @@
-<div fxLayout="row" fxLayoutAlign="space-evenly">
-  <div fxFlex="20"><app-user-info [profile]="profile"></app-user-info></div>
-  <div fxFlex="80"><app-user-tabs [profile]="profile"></app-user-tabs></div>
+<div class="flex-container" fxLayoutAlign="space-evenly">
+  <div class="item info" ><app-user-info [profile]="profile"></app-user-info></div>
+  <div class="item tabs"><app-user-tabs [profile]="profile"></app-user-tabs></div>
 </div>
 
 <!-- <div fxLayout="row" fxLayout.xs="column" fxLayoutWrap fxLayoutGap="0.5%" fxLayoutAlign="center">
