+++ SocialNetwork/client/src/app/components/profile-page/user-page/user-page.component.scss
@@ -1,3 +1,17 @@
-mat-grid-tile {
-    background: lightblue;
+  .flex-container {
+    resize: horizontal;
+    overflow: hidden;
+    flex-direction: row;
+    display: flex;
+    margin: 1em;
+  }
+
+  .info {
+    flex: 1;
+   // max-width: 230px;
+  }
+
+  .tabs {
+    flex: 3;
+    flex-grow: 10;
   }
