+++ SocialNetwork/client/src/app/components/registration-form/registration-form.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { User } from '../../models/user';
-import { InputDataValidatorService } from '../../services/validators/input-data-validator.service';
+import { InputDataValidatorService } from '../../validators/input-data-validator.service';
