+++ SocialNetwork/client/src/app/validators/input-data-validator.service.spec.ts
