+++ SocialNetwork/client/src/app/validators/input-data-validator.service.ts
